Fix the user's creation error when first connecting and add a new attribute form for the user's fullname
When attempting to add a user to the database who is try to authentificate with the AuthCAS plugin :
Internal Server Error
CDbCommand failed to execute the SQL statement: SQLSTATE [HY000]: General error: 1364 Field 'entity_id' does not have a default value
An internal error occurred while the Web server was processing your request. Please contact the webmaster to report this problem.